Update Step-Drugs -
= > Update two lists of radio buttons adding a value=0 for each 
Please do make sure that does button (value=0) are the default value when "adding a pet". Otherwise we keep the same logic and we pre-filled everything with the previous state of the drug_history. (as it is perfectly now). I just want that the 0 Value radiobutton are checked by default when adding a pet.